pi-code/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
anthropic==0.34.0
Pillow==10.4.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic==2.9.2
python-dotenv==1.0.1
